+++ crates/world_state/src/behavior/tree.rs
@@ -7,14 +7,13 @@ use crate::{
         behavior_tree::Node,
         condition::{
             has_ball_position, is_ball_interception_candidate, is_close_to_ball,
-            is_closest_to_ball, is_fallen, is_goalkeeper, is_in_sub_state, is_primary_state,
-            is_remote_controlled,
+            is_closest_to_ball, is_fallen, is_goalkeeper, is_primary_state, is_remote_controlled,
         },
         head::{look_at_ball_subtree, look_straight_ahead, search_for_lost_ball_subtree},
         kick::{intercept, kick_subtree},
         node::Blackboard,
         search::{has_suggested_search_position, leuchtturm, walk_to_search_position},
-        substates::sub_state_subtree,
+        substates::{is_in_sub_state, sub_state_subtree},
         switch_motion_type::switch_motion_type,
         walk::{walk_alternatives_subtree, walk_to_ball_subtree},
     },
